requirements.txt
streamlit==1.17.0
pandas
numpy
matplotlib
seaborn
scikit-learn==1.1.3
Pillow
altair==4.0
xgboost==1.7.1